Fixed duct detectors not being placed, added some post-school commands

+++ src/main/java/com/feed_the_beast/mods/ftbacademymod/blocks/BlockDuctDetector.java
@@ -156,7 +156,7 @@ public class BlockDuctDetector extends BlockDetectorBase
 						{
 							NBTTagCompound nbt = attachments.getCompoundTagAt(0);
 
-							if (nbt.getInteger("Flags") == variant.flags)
+							if (nbt.getByte("Flags") == variant.flags)
 							{
 								HashSet<FilterItem> inv = new HashSet<>();
